+++ model_multigpu.py
@@ -29,15 +29,15 @@ class Model(object):
 
     def _input_layer(self):
         with tf.variable_scope('Train_Inputs'):
-            self.train_query = tf.placeholder(dtype=tf.int32, shape=(None, self.MAX_QUERY_WORD_SIZE), name='query')
-            self.train_title = tf.placeholder(dtype=tf.int32, shape=(None, self.MAX_TITLE_WORD_SIZE), name='title')
-            self.train_labels = tf.placeholder(dtype=tf.int32, shape=(None, ), name='labels')
+            self.train_query = tf.placeholder(dtype=tf.int64, shape=(None, self.MAX_QUERY_WORD_SIZE), name='query')
+            self.train_title = tf.placeholder(dtype=tf.int64, shape=(None, self.MAX_TITLE_WORD_SIZE), name='title')
+            self.train_labels = tf.placeholder(dtype=tf.int64, shape=(None, ), name='labels')
 
         with tf.variable_scope('Predict_Inputs'):
-            self.predict_query = tf.placeholder(dtype=tf.int32, shape=(None, self.MAX_QUERY_WORD_SIZE), name='query')
-            self.predict_title = tf.placeholder(dtype=tf.int32, shape=(None, self.MAX_TITLE_WORD_SIZE), name='title')
-            self.predict_labels = tf.placeholder(dtype=tf.int32, shape=(None,), name='labels')
-            self.title_num = tf.placeholder(dtype=tf.int32, shape=(1,),  name='title_num')
+            self.predict_query = tf.placeholder(dtype=tf.int64, shape=(None, self.MAX_QUERY_WORD_SIZE), name='query')
+            self.predict_title = tf.placeholder(dtype=tf.int64, shape=(None, self.MAX_TITLE_WORD_SIZE), name='title')
+            self.predict_labels = tf.placeholder(dtype=tf.int64, shape=(None,), name='labels')
+            self.title_num = tf.placeholder(dtype=tf.int64, shape=(1,),  name='title_num')
 
     def _embed_layer(self, query, title):
         with tf.variable_scope('Embedding_layer'), tf.device("/cpu:0"):
@@ -166,7 +166,7 @@ class Model(object):
 
             self.losses = tf.maximum(0.0, tf.subtract(FLAGS.margin, tf.subtract(self.score_pos, self.score_neg)))
             self.loss = tf.reduce_mean(self.losses)
-            # self.loss = tf.reduce_mean(tf.log(1.0 + tf.exp(- 2.0 * self.sub)))
+            #self.loss = tf.reduce_mean(tf.log(1.0 + tf.exp(- 2.0 * self.sub)))
             self.sm_loss_op = tf.summary.scalar('Loss', self.loss)
